Article Module Structure Finished

+++ src/main/java/com/cloud/project/dao/UserDao.java
@@ -3,7 +3,7 @@ package com.cloud.project.dao;
 import com.cloud.project.entity.User;
 
 public interface UserDao {
-    User findById(Long id);
+    User findById(int id);
     User findByUsername(String username);
     int insert(User user);
 }
